microservice/requirements.txt
fastapi
uvicorn
pydantic
PyJWT
passlib[bcrypt]
bcrypt==3.2.0
cryptography
python-multipart
